Update wc_get_page_id compatability return type
The WooCommerce core version of wc_get_page_id always returns an integer
and not a string. See here:

https://github.com/woocommerce/woocommerce/blob/9c83a07145ea5c451adc489f21eeaf895825b683/plugins/woocommerce/includes/wc-page-functions.php#L36-L57

Ensuring that this compatibility version of the function also returns an integer could prevent 
subtle issues with themes or plugins that expect `wc_get_page_id` to always return an 
integer. 

In WC core a return value of -1 means the page was not found, which seems to match the
desired behavior here.