Enhance channel name handling by adding sanitization methods (#486)
* Enhance channel name handling by adding sanitization methods

- Implemented `sanitize_channel_name` method to sanitize channel names before saving or using them.
- Updated the settings registration to include a sanitize callback.
- Modified the default value for the channel name to use the sanitized version of the home URL.
- Overrode `render_field` method to ensure the channel name is sanitized during rendering.

These changes ensure that the default channel name does not include periods which leads to an error creating the channel

* Spacing

Proper spacing for wp standards

* Remove regex
Remove the unnecessary regex checking for the channel name. This PR should just be replacing the periods. There are API-level sanitizations that prevent these characters from being included in channel names.